feat(ui): implement table cards for constraint configuration
