+++ frontend/src/app/dashboard/[scenarioId]/page.tsx
@@ -1,6 +1,6 @@
 import type { UserScenarioId } from "@/convex/types";
 
-import { LiveScenarioUpdateCard } from "@/features/scenario/update";
+import { LiveScenarioUpdate } from "@/features/scenario/update";
 
 interface Params {
   scenarioId: UserScenarioId;
@@ -12,9 +12,5 @@ interface PageProps {
 
 export default async function Page({ params }: PageProps) {
   const { scenarioId } = await params;
-  return (
-    <div className="mx-auto w-full max-w-7xl justify-center space-y-8 px-6 py-8">
-      <LiveScenarioUpdateCard id={scenarioId} />
-    </div>
-  );
+  return <LiveScenarioUpdate id={scenarioId} />;
 }
